fix conflict

+++ SpringBootAPI/HRMS_SWP490_G2/src/main/java/com/swp490_g2/hrms/service/RestaurantService.java
@@ -1,12 +1,10 @@
 package com.swp490_g2.hrms.service;
 
 import com.swp490_g2.hrms.common.utils.CommonUtils;
-import com.swp490_g2.hrms.entity.File;
-import com.swp490_g2.hrms.entity.Restaurant;
-import com.swp490_g2.hrms.entity.User;
+import com.swp490_g2.hrms.entity.*;
 import com.swp490_g2.hrms.repositories.RestaurantRepository;
-import com.swp490_g2.hrms.requests.RestaurantInformationRequest;
 import com.swp490_g2.hrms.requests.SearchRestaurantsRequest;
+import com.swp490_g2.hrms.requests.RestaurantInformationRequest;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -15,10 +13,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 @Service
 @Getter
@@ -78,6 +73,7 @@ public class RestaurantService {
     }
 
     public void update(Restaurant restaurant) {
+        restaurant.setAddress(addressService.populateLatLng(restaurant.getAddress()));
         restaurantRepository.save(restaurant);
     }
 
@@ -144,11 +140,16 @@ public class RestaurantService {
     }
 
     public void deleteRestaurantById(Long restaurantId) {
-        Restaurant restaurant = restaurantRepository.getById(restaurantId);
-        User user = userService.getByRestaurantId(restaurantId);
-        RestaurantInformationRequest request = RestaurantInformationRequest.set(user, restaurant);
+        RestaurantInformationRequest request = getRestaurantById(restaurantId);
         if(Objects.nonNull(request)) {
-            int n = restaurantRepository.deleteSellerRestaurant(restaurantId, request.getUser().getId());
+            if(request.getRestaurantCategory() != null || request.getRestaurantCategory().size() > 0) {
+                for (RestaurantCategory category: request.getRestaurantCategory()) {
+                    restaurantRepository.deleteRestaurantCategory(restaurantId, category.getId());
+                }
+            }
+            if(Objects.nonNull(request.getUser())) {
+                restaurantRepository.deleteSellerRestaurant(restaurantId, request.getUser().getId());
+            }
             restaurantRepository.deleteById(restaurantId);
         }
     }
